# awesomeProject1/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: ToDo List API
  version: 1.0.0
  description: API for managing users and tasks
servers:
  - url: http://localhost:8084
    description: Local development server
paths:
  /tasks:
    get:
      operationId: GetTasks
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of all tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      operationId: PostTasks
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}:
    get:
      operationId: GetTasksById
      summary: Get task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the task
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: PatchTasksById
      summary: Update a task
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskBody'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: DeleteTasksById
      summary: Delete a task
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the task
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/tasks:
    get:
      operationId: GetUsersIdTasks
      summary: Get all tasks for specific user
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user
      responses:
        '200':
          description: A list of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          readOnly: true
        task:
          type: string
          example: "Buy groceries"
        is_done:
          type: boolean
          example: false
        userId:
          type: integer
          format: int64
          example: 1
      required:
        - task
        - userId
    TaskCreate:
      type: object
      properties:
        task:
          type: string
          example: "Complete project"
        is_done:
          type: boolean
          example: false
        user_id:
          type: integer
          format: int64
          example: 1
      required:
        - task
        - user_id
    TaskBody:
      type: object
      properties:
        task:
          type: string
          example: "Updated task description"
        is_done:
          type: boolean
          example: true
      required:
        - task
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Error description"
        code:
          type: integer
          example: 404
        details:
          type: string
          example: "Additional error details"
      required:
        - message
        - code